\name{export.default}
\alias{export.default}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{export.default method }
\description{
Default method \code{export}
}
\usage{
export.default(x, filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object to export}
  \item{filename}{ target filename }
}
\value{
  None.
}
\seealso{ \code{\link{export.listpoly}}}
\keyword{IO}
