\name{export}
\alias{export}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Generic for 'export'}
\description{
 Generic function for 'export'
}
\usage{
export(x, filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ object to export}
  \item{filename}{ target filename }
}
\value{
  None.
}
\seealso{ \code{\link{export.listpoly}}}
\keyword{IO}
