\name{listpoly-class}
\docType{class}
\alias{listpoly-class}

\title{Class 'listpoly'  }
\description{ A class to store a suite of polygons and its methods.}
\section{Objects from the Class}{
  Objects can be created by calls of the function
  \code{\link{crlistpoly}} and  \code{\link{generPoly}} }
\section{Slots}{
	 \describe{
	   \item{\code{}}{List where each component is an object of
	    \code{\link{poly-class}}
  }
}
}
\section{Methods}{
\code{\link{plot.listpoly}}, 
\code{\link{range.listpoly}},
\code{\link{export.listpoly}}}
\seealso{
  \code{\link{poly-class}}
}
\keyword{classes}
