\name{plot.poly}
\alias{plot.poly}
\title{ Plot of an object of class 'poly' }
\description{
  Plot an object  of class \code{"\link[=poly-class]{poly}"} via the function \code{\link[splancs]{polymap}} of the library
  \code{splancs}.
  }
\usage{
\method{plot}{poly}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ Object of class \code{"\link[=poly-class]{poly}"}. }
  \item{\dots}{ Arguments passed as-it to 'polymap'.}
}
\value{
  None.
}
\keyword{hplot}
