\name{poly-class}
\docType{class}
\alias{poly-class}

\title{Class 'poly'  }
\description{ A class to store the coordinates of a polygons and its methods.}
\section{Objects from the Class}{
  Objects can be created by calls of the functions
  \code{\link{as.poly}} and \code{\link{crpoly}} }
\section{Slots}{
	 \describe{
	   \item{m}{Matrix with two columns \code{xcoord}
	     and \code{ycoord}}
  }
}
\section{Methods}{
\code{\link{plot.poly}}
}
\seealso{
  \code{\link{listpoly-class}}
}
\keyword{classes}
