\name{RCALI-package}
\alias{RCALI-package}
\docType{package}
\title{
Calculation of the Integrated Flow of Particles Between Polygons
}
\description{
 Calculate the flow of particles between polygons by two integration methods: integration by a cubature method and integration on a grid of points.
}
\details{
\tabular{ll}{
Depends: \tab splancs\cr
License: \tab GPL version 2 or later\cr
This package is an interface between R and the programme CaliFloPP: 
see: \tab \url{http://www.inra.fr/miaj/public/logiciels/califlopp}\cr
 News\tab             \code{file.show(system.file("NEWS", package="RCALI"))}
}
Main function: \code{\link{califlopp}}
}
\author{
Annie Bouvier <annie.bouvier@jouy.inra.fr>
}
\keyword{ package }
