\name{RCALI-package}
\alias{RCALI-package}
\alias{RCALI}
\docType{package}
\title{
Calculation of the Integrated Flow of Particles Between Polygons
}
\description{
 Calculate the flow of particles between polygons by two integration methods: integration by a cubature method and integration on a grid of points.
}
\details{
\tabular{ll}{
Depends: \tab splancs\cr
License: \tab GPL version 2 or later\cr
This package is an interface between R and the programme CaliFloPP: \cr
\url{http://genome.jouy.inra.fr/logiciels/califlopp}\cr
 News\tab             \code{file.show(system.file("NEWS", package="RCALI"))}\cr
Main function: \tab \code{\link{califlopp}}
}
}
\author{
Annie Bouvier <annie.bouvier@inra.fr>
}
\keyword{ package }
